More review stuff

+++ default/python/AI/freeorion_tools/bonus_calculation.py
@@ -12,7 +12,7 @@ class Bonus(NamedTuple):
         return self.value if stability >= self.required_stability else 0.0
 
 
-def adjust_direction(threshold: float, current: float, target) -> float:
+def adjust_direction(threshold: float, current: float, target) -> int:
     """
     Returns -1, 0 or 1, depending on whether compared to current, target drops below, is on the same side or
     goes above the threshold.
